+++ src/main/java/net/dv8tion/jda/core/entities/Role.java
@@ -82,13 +82,13 @@ public interface Role extends ISnowflake, IMentionable, Comparable<Role>
     boolean isMentionable();
 
     /**
-     * The <code>int</code> representation of the literal permissions that this {@link net.dv8tion.jda.core.entities.Role Role} has.<br>
+     * The <code>long</code> representation of the literal permissions that this {@link net.dv8tion.jda.core.entities.Role Role} has.<br>
      * <b>NOTE:</b> these do not necessarily represent the permissions this role will have in a {@link net.dv8tion.jda.core.entities.Channel Channel}.
      *
      * @return
-     *      Never-negative int containing offset permissions of this role.
+     *      Never-negative long containing offset permissions of this role.
      */
-    int getPermissionsRaw();
+    long getPermissionsRaw();
 
     /**
      * A list of the literal {@link net.dv8tion.jda.core.Permission Permissions} that this {@link net.dv8tion.jda.core.entities.Role Role} has.<br>
